+++ src/js/leaflet.storage.layer.js
@@ -64,7 +64,7 @@ L.Storage.DataLayer = L.Class.extend({
         if (this.isClustered()) {
             this.layer = new L.MarkerClusterGroup({
                 polygonOptions: {
-                    color: this.options.color || this.map.getDefaultOption('color')
+                    color: this.getColor()
                 },
                 iconCreateFunction: function (cluster) {
                     return new L.Storage.Icon.Cluster(self, cluster);
@@ -353,6 +353,10 @@ L.Storage.DataLayer = L.Class.extend({
         return new L.Storage.Icon[this.getIconClass()](this.map);
     },
 
+    getColor: function () {
+        return this.options.color || this.map.getDefaultOption('color');
+    },
+
     getDeleteUrl: function () {
         return L.Util.template(this.map.options.urls.datalayer_delete, {'pk': this.storage_id, 'map_id': this.map.options.storage_id});
 
@@ -436,7 +440,7 @@ L.Storage.DataLayer = L.Class.extend({
                     this.resetLayer();
                 }
                 if (field === "options.color" && this.isClustered()) {
-                    this.layer.options.polygonOptions.color = this.options.color || this.map.getDefaultOption('color');
+                    this.layer.options.polygonOptions.color = this.getColor();
                 }
                 this.display();
             }
@@ -594,6 +598,18 @@ L.Storage.DataLayer = L.Class.extend({
 
     getName: function () {
         return this.options.name || L._('Untitled layer');
+    },
+
+    renderToolbox: function (container) {
+        var toggle = L.DomUtil.create('i', 'layer-toggle', container),
+            zoom_to = L.DomUtil.create('i', 'layer-zoom_to', container),
+            edit = L.DomUtil.create('i', "layer-edit show-on-edit", container);
+        zoom_to.title = L._('Zoom to layer extent');
+        toggle.title = L._('Show/hide layer');
+        edit.title = L._('Edit');
+        L.DomEvent.on(toggle, 'click', this.toggle, this);
+        L.DomEvent.on(zoom_to, 'click', this.zoomTo, this);
+        L.DomEvent.on(edit, 'click', this.edit, this);
     }
 
 });
@@ -606,7 +622,8 @@ L.TileLayer.include({
             maxZoom: this.options.maxZoom,
             attribution: this.options.attribution,
             url_template: this._url,
-            name: this.options.name
+            name: this.options.name,
+            tms: this.options.tms
         };
     },